invoice_extractor/requirements.txt
pdfplumber==0.11.9
camelot-py[cv]==1.0.9
pandas==3.0.0
pydantic==2.10.4
pyyaml==6.0.2
openpyxl==3.1.5
